requirements.txt
streamlit>=1.35.0
groq>=0.11.0
# RAG stack
langchain>=0.2.0
langchain-community>=0.2.0
langchain-groq>=0.1.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
# Web search
tavily-python>=0.3.0
# Scraping for RAG ingestion
requests>=2.31.0
beautifulsoup4>=4.12.0
# Auth & persistence
supabase>=2.4.0
# Token counting
tiktoken>=0.7.0
# Logging & monitoring
loguru>=0.7.0
# Utilities
python-dotenv>=1.0.0
pydantic>=2.6.0
